+++ src/main/java/backend/github/UpdateIssuesTask.java
@@ -6,6 +6,7 @@ import backend.resource.Model;
 import backend.resource.TurboIssue;
 import org.apache.commons.lang3.tuple.ImmutableTriple;
 import org.apache.logging.log4j.Logger;
+import org.eclipse.egit.github.core.PullRequest;
 import util.HTLog;
 
 import java.util.Date;
@@ -26,16 +27,21 @@ public class UpdateIssuesTask extends GitHubRepoTask<GitHubRepoTask.Result<Turbo
     public void run() {
         ImmutableTriple<List<TurboIssue>, String, Date> changes = repo.getUpdatedIssues(model.getRepoId(),
             model.getUpdateSignature().issuesETag, model.getUpdateSignature().lastCheckTime);
+        List<PullRequest> updatedPullRequests = repo.getUpdatedPullRequests(
+                model.getRepoId(), model.getUpdateSignature().lastCheckTime);
 
         List<TurboIssue> existing = model.getIssues();
-        List<TurboIssue> changed = changes.left;
+        List<TurboIssue> updatedIssues = changes.left;
 
         logger.info(HTLog.format(model.getRepoId(), "%s issue(s)) changed%s",
-            changed.size(), changed.isEmpty() ? "" : ": " + changed));
+                updatedIssues.size(), updatedIssues.isEmpty() ? "" : ": " + updatedIssues));
+        logger.info(HTLog.format(model.getRepoId(), "%s pr(s)) changed%s",
+                updatedPullRequests.size(), updatedPullRequests.isEmpty() ? "" : ": " + updatedPullRequests));
 
-        List<TurboIssue> updated = changed.isEmpty()
+        List<TurboIssue> updated = updatedIssues.isEmpty()
             ? existing
-            : TurboIssue.reconcile(model.getRepoId(), existing, changed);
+            : TurboIssue.reconcile(model.getRepoId(), existing, updatedIssues);
+        updated = TurboIssue.combineWithPullRequests(updated, updatedPullRequests);
 
         response.complete(new Result<>(updated, changes.middle, changes.right));
     }